requirements.txt
asgiref==3.10.0
charset-normalizer==3.4.4
Django==5.2.7
djangorestframework==3.16.1
dotenv==0.9.9
gunicorn==23.0.0
packaging==25.0
pillow==12.0.0
python-dotenv==1.2.1
pytz==2025.2
reportlab==4.4.4
sqlparse==0.5.3
whitenoise==6.11.0